feat(og): reusable dynamic OG template + per-component cards
Phase 1 of the SEO/AIO ("Discovery", N11) work: make previews dynamic so a
shared link picks up the page's own title + SEO description + the site icon,
not one static image.

- lib/og/mzizi-og.tsx: extract the honeycomb "hive" card into renderMziziOg
  ({ title, eyebrow, description }) — the single template every route reuses.
  Title size scales with length; hero is the site icon (nyuchi bee); mineral
  accents + faint hex lattice from the design DB via paletteColor().
- app/opengraph-image.tsx: root card now just calls the template.
- app/components/[name]/opengraph-image.tsx: per-component card pulling the
  component's name + description + layer from the registry, with a graceful
  slug fallback when the DB is unreachable.

This is the artifact the Mzizi toolkit (skill + CLI scaffolder + MCP) will
ship so any consumer app guarantees dynamic previews.
